requirements.txt
Django==4.2.21
psycopg2-binary==2.9.9
python-decouple==3.8
django-crispy-forms==2.1
crispy-tailwind==0.5.0
plotly==5.17.0
pandas==2.1.4
numpy==1.24.4
django-extensions==3.2.3
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0 